break up tests to support category use
