+++ subs-api/src/test/java/uk/ac/ebi/subs/api/ApiIntegrationTest.java
@@ -11,20 +11,15 @@ import org.json.JSONObject;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.embedded.LocalServerPort;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.hateoas.Link;
-import org.springframework.hateoas.MediaTypes;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import uk.ac.ebi.subs.ApiApplication;
-import uk.ac.ebi.subs.RabbitMQDependentTest;
 import uk.ac.ebi.subs.data.Submission;
 import uk.ac.ebi.subs.data.client.Sample;
 import uk.ac.ebi.subs.repository.repos.SubmissionRepository;
@@ -34,13 +29,13 @@ import uk.ac.ebi.subs.repository.repos.status.SubmissionStatusRepository;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.assertThat;
+
+import static uk.ac.ebi.subs.api.ApiIntegrationTestHelper.*;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = ApiApplication.class, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
@@ -52,6 +47,9 @@ public class ApiIntegrationTest {
     private int port;
     private String rootUri;
 
+
+
+    private ApiIntegrationTestHelper testHelper;
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -68,6 +66,7 @@ public class ApiIntegrationTest {
 
     @Before
     public void buildUp() throws URISyntaxException {
+
         rootUri = "http://localhost:" + port + "/api";
         submissionRepository.deleteAll();
         sampleRepository.deleteAll();
@@ -90,6 +89,8 @@ public class ApiIntegrationTest {
                 }
             }
         });
+
+        testHelper = new ApiIntegrationTestHelper(objectMapper,rootUri);
     }
 
     @After
@@ -102,17 +103,17 @@ public class ApiIntegrationTest {
 
     @Test
     public void checkRootRels() throws UnirestException, IOException {
-        Map<String, String> rootRels = rootRels();
+        Map<String, String> rootRels = testHelper.rootRels();
 
         assertThat(rootRels.keySet(), hasItems("submissions", "samples"));
     }
 
     @Test
     public void postSubmission() throws UnirestException, IOException {
-        Map<String, String> rootRels = rootRels();
+        Map<String, String> rootRels = testHelper.rootRels();
 
         Submission submission = Helpers.generateSubmission();
-        HttpResponse<JsonNode> submissionResponse = postSubmission(rootRels, submission);
+        HttpResponse<JsonNode> submissionResponse = testHelper.postSubmission(rootRels, submission);
 
         List<SubmissionStatus> submissionStatuses = submissionStatusRepository.findAll();
         assertThat(submissionStatuses, notNullValue());
@@ -122,122 +123,18 @@ public class ApiIntegrationTest {
         assertThat(submissionStatus.getStatus(), equalTo("Draft"));
     }
 
-    @Test
-    @Category(RabbitMQDependentTest.class)
-    //Requires dispatcher to delete the contents
-    public void postThenDeleteSubmission() throws UnirestException, IOException {
-        Map<String, String> rootRels = rootRels();
 
-        String submissionLocation = submissionWithSamples(rootRels);
-        HttpResponse<JsonNode> deleteResponse = Unirest.delete(submissionLocation)
-                .headers(standardPostHeaders())
-                .asJson();
-
-        assertThat(deleteResponse.getStatus(), equalTo(HttpStatus.NO_CONTENT.value()));
-
-        List<uk.ac.ebi.subs.repository.model.Submission> submissions = submissionRepository.findAll();
-        assertThat(submissions, empty());
-
-    }
-
-    /**
-     * create a submission with some samples and submit it
-     *
-     * @throws IOException
-     * @throws UnirestException
-     */
-    @Test
-    @Category(RabbitMQDependentTest.class)
-    public void simpleSubmissionWorkflow() throws IOException, UnirestException {
-        Map<String, String> rootRels = rootRels();
-
-        String submissionLocation = submissionWithSamples(rootRels);
-
-        HttpResponse<JsonNode> submissionGetResponse = Unirest
-                .get(submissionLocation)
-                .headers(standardGetHeaders())
-                .asJson();
-
-        assertThat(submissionGetResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
-        JSONObject payload = submissionGetResponse.getBody().getObject();
-
-        Map<String,String> rels = relsFromPayload(payload);
-
-        assertThat(rels.get("submissionStatus"),notNullValue());
-        String submissionStatusLocation = rels.get("submissionStatus");
-
-        HttpResponse<JsonNode> submissionStatusGetResponse = Unirest
-                .get(submissionStatusLocation)
-                .headers(standardGetHeaders())
-                .asJson();
-
-        assertThat(submissionStatusGetResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
-        JSONObject statusPayload = submissionStatusGetResponse.getBody().getObject();
-
-        rels = relsFromPayload(statusPayload);
-
-        assertThat(rels.get("self"),notNullValue());
-        submissionStatusLocation = rels.get("self");
-
-
-        //update the submission
-        //create a new submission
-        HttpResponse<JsonNode> submissionPatchResponse = Unirest.patch(submissionStatusLocation)
-                .headers(standardPostHeaders())
-                .body("{\"status\": \"Submitted\"}")
-                .asJson();
-
-
-        assertThat(submissionPatchResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
-    }
 
     @Test
     public void submissionWithSamples() throws IOException, UnirestException {
-        Map<String, String> rootRels = rootRels();
+        Map<String, String> rootRels = testHelper.rootRels();
 
-        String submissionLocation = submissionWithSamples(rootRels);
+        String submissionLocation = testHelper.submissionWithSamples(rootRels);
 
 
     }
 
-    private String submissionWithSamples(Map<String, String> rootRels) throws UnirestException, IOException {
-        Submission submission = Helpers.generateSubmission();
-        HttpResponse<JsonNode> submissionResponse = postSubmission(rootRels, submission);
 
-        String submissionLocation = submissionResponse.getHeaders().get("Location").get(0).toString();
-        Map<String, String> submissionRels = relsFromPayload(submissionResponse.getBody().getObject());
-
-        assertThat(submissionRels.get("samples"), notNullValue());
-
-        List<Sample> testSamples = Helpers.generateTestSamples();
-        //add samples to the submission
-        for (Sample sample : testSamples) {
-
-            sample.setSubmission(submissionLocation);
-
-            HttpResponse<JsonNode> sampleResponse = Unirest.post(rootRels.get("samples"))
-                    .headers(standardPostHeaders())
-                    .body(sample)
-                    .asJson();
-
-            assertThat(sampleResponse.getStatus(), is(equalTo(HttpStatus.CREATED.value())));
-        }
-
-        //retrieve the samples
-        String submissionSamplesUrl = submissionRels.get("samples");
-
-        HttpResponse<JsonNode> samplesQueryResponse = Unirest.get(submissionSamplesUrl)
-                .headers(standardGetHeaders())
-                .asJson();
-
-        assertThat(samplesQueryResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
-
-        JSONObject payload = samplesQueryResponse.getBody().getObject();
-        JSONArray sampleList = payload.getJSONObject("_embedded").getJSONArray("samples");
-
-        assertThat(sampleList.length(), is(equalTo(testSamples.size())));
-        return submissionLocation;
-    }
 
     /**
      * Make multiple submissions with the same contents. Use the sample history endpoint to check that you can
@@ -248,7 +145,7 @@ public class ApiIntegrationTest {
      */
     @Test
     public void sampleVersions() throws IOException, UnirestException {
-        Map<String, String> rootRels = rootRels();
+        Map<String, String> rootRels = testHelper.rootRels();
 
 
         int numberOfSubmissions = 5;
@@ -257,10 +154,10 @@ public class ApiIntegrationTest {
         List<Sample> testSamples = Helpers.generateTestSamples();
 
         for (int i = 0; i < numberOfSubmissions; i++) {
-            HttpResponse<JsonNode> submissionResponse = postSubmission(rootRels, submission);
+            HttpResponse<JsonNode> submissionResponse = testHelper.postSubmission(rootRels, submission);
 
             String submissionLocation = submissionResponse.getHeaders().get("Location").get(0).toString();
-            Map<String, String> submissionRels = relsFromPayload(submissionResponse.getBody().getObject());
+            Map<String, String> submissionRels = testHelper.relsFromPayload(submissionResponse.getBody().getObject());
 
             assertThat(submissionRels.get("samples"), notNullValue());
 
@@ -285,7 +182,7 @@ public class ApiIntegrationTest {
         assertThat(domainQueryResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
 
         JSONObject domainPayload = domainQueryResponse.getBody().getObject();
-        Map<String, String> domainRels = relsFromPayload(domainPayload);
+        Map<String, String> domainRels = testHelper.relsFromPayload(domainPayload);
 
         String domainSamplesUrl = domainRels.get("samples");
 
@@ -301,13 +198,13 @@ public class ApiIntegrationTest {
         for (int i = 0; i < domainSamples.length(); i++){
             JSONObject domainSample = domainSamples.getJSONObject(i);
 
-            Map<String,String> sampleRels = relsFromPayload(domainSample);
+            Map<String,String> sampleRels = testHelper.relsFromPayload(domainSample);
             String selfUrl = sampleRels.get("self");
 
             HttpResponse<JsonNode> sampleResponse = Unirest.get(selfUrl).headers(standardGetHeaders()).asJson();
             assertThat(sampleResponse.getStatus(), is(equalTo(HttpStatus.OK.value())));
             JSONObject samplePayload = sampleResponse.getBody().getObject();
-            sampleRels = relsFromPayload(samplePayload);
+            sampleRels = testHelper.relsFromPayload(samplePayload);
 
             String historyUrl = sampleRels.get("history");
 
@@ -327,27 +224,17 @@ public class ApiIntegrationTest {
     }
 
 
-    private HttpResponse<JsonNode> postSubmission(Map<String, String> rootRels, Submission submission) throws UnirestException {
-        //create a new submission
-        HttpResponse<JsonNode> submissionResponse = Unirest.post(rootRels.get("submissions"))
-                .headers(standardPostHeaders())
-                .body(submission)
-                .asJson();
 
-        assertThat(submissionResponse.getStatus(), is(equalTo(HttpStatus.CREATED.value())));
-        assertThat(submissionResponse.getHeaders().get("Location"), notNullValue());
-        return submissionResponse;
-    }
 
     @Test
     public void testPut() throws IOException, UnirestException {
-        Map<String, String> rootRels = rootRels();
+        Map<String, String> rootRels = testHelper.rootRels();
 
         Submission submission = Helpers.generateSubmission();
-        HttpResponse<JsonNode> submissionResponse = postSubmission(rootRels, submission);
+        HttpResponse<JsonNode> submissionResponse = testHelper.postSubmission(rootRels, submission);
 
         String submissionLocation = submissionResponse.getHeaders().getFirst("Location");
-        Map<String, String> submissionRels = relsFromPayload(submissionResponse.getBody().getObject());
+        Map<String, String> submissionRels = testHelper.relsFromPayload(submissionResponse.getBody().getObject());
 
         assertThat(submissionRels.get("samples"), notNullValue());
 
@@ -370,7 +257,7 @@ public class ApiIntegrationTest {
         sample.setSubmission(submissionLocation);
 
         HttpResponse<JsonNode> samplePutResponse = Unirest.put(sampleLocation)
-                .headers(standardPostHeaders())
+                .headers(ApiIntegrationTestHelper.standardPostHeaders())
                 .body(sample)
                 .asJson();
 
@@ -381,52 +268,10 @@ public class ApiIntegrationTest {
     }
 
 
-    private Map<String, String> standardGetHeaders() {
-        Map<String, String> h = new HashMap<>();
-        h.put("accept", MediaTypes.HAL_JSON_VALUE);
-        return h;
-    }
-
-    private Map<String, String> standardPostHeaders() {
-        Map<String, String> h = new HashMap<>();
-        h.put("accept", MediaTypes.HAL_JSON_VALUE);
-        h.put("Content-Type", MediaType.APPLICATION_JSON_VALUE);
-        return h;
-    }
 
 
-    private Map<String, String> rootRels() throws UnirestException, IOException {
-        HttpResponse<JsonNode> response = Unirest.get(rootUri)
-                .headers(standardGetHeaders())
-                .asJson();
-
-        assertThat(response.getStatus(), is(equalTo(HttpStatus.OK.value())));
-        JSONObject payload = response.getBody().getObject();
-
-        return relsFromPayload(payload);
-    }
-
-    private Map<String, String> relsFromPayload(JSONObject payload) throws IOException {
-        assertThat((Set<String>) payload.keySet(), hasItem("_links"));
-
-        JSONObject links = payload.getJSONObject("_links");
-
-        Map<String, String> rels = new HashMap<>();
 
 
-        for (Object key : links.keySet()) {
-
-            assertThat(key.getClass(), typeCompatibleWith(String.class));
-
-            Object linkJson = links.get(key.toString());
-            Link link = objectMapper.readValue(linkJson.toString(), Link.class);
-            String href = link.withSelfRel().expand().getHref();
-
-            rels.put((String) key, href);
-
-        }
-        return rels;
-    }
 
 
 }
